# openapi.yaml
openapi: 3.0.3
info:
  title: APF Portal Authentication API
  description: |
    Secure authentication and authorization API for the Accountants and Procurement Professionals portal.
    
    Features:
    - Two-factor authentication via email OTP
    - JWT-based session management
    - Role-based access control (Admin/Member)
    - Rate limiting and audit logging
    - Password reset functionality
    
    ## Authentication Flow
    1. POST /login with email/password → receive session_id
    2. POST /verify-otp with session_id and OTP → receive JWT tokens
    3. Use access_token in Authorization header for protected endpoints
    4. Refresh access_token using refresh_token when expired
  version: 1.0.0
  contact:
    email: support@apfportal.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/auth
    description: Local development server
  - url: https://api.apfportal.com/api/auth
    description: Production server
...
paths:
...
  /refresh:
    post:
      tags:
        - Token Management
      summary: Refresh access token
      description: |
        Generate new access token using valid refresh token.
        Access tokens expire after 1 hour.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/InvalidToken'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /verify-otp endpoint
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            message:
              type: string
              example: Error description
            details:
              type: object
              additionalProperties: true
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Email and password are required
...
    InvalidToken:
      description: Invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: INVALID_TOKEN
              message: Invalid or expired refresh token
...